move utility modules to src/echodataflow/utils
